requirements.txt
streamlit
python-pptx
pypdf